move code to more organized package structure for services and models
