+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/models/productlist/ProductList.java
@@ -23,7 +23,7 @@ import org.osgi.annotation.versioning.ProviderType;
 
 import com.adobe.cq.commerce.core.components.models.common.ProductListItem;
 import com.adobe.cq.commerce.core.components.models.retriever.AbstractCategoryRetriever;
-import com.adobe.cq.commerce.core.search.SearchResultsSet;
+import com.adobe.cq.commerce.core.search.models.SearchResultsSet;
 
 @ProviderType
 public interface ProductList {
