+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/components/models/searchresults/SearchResults.java
@@ -21,8 +21,8 @@ import javax.annotation.Nonnull;
 import org.osgi.annotation.versioning.ProviderType;
 
 import com.adobe.cq.commerce.core.components.models.common.ProductListItem;
-import com.adobe.cq.commerce.core.search.SearchAggregation;
-import com.adobe.cq.commerce.core.search.SearchResultsSet;
+import com.adobe.cq.commerce.core.search.models.SearchAggregation;
+import com.adobe.cq.commerce.core.search.models.SearchResultsSet;
 
 /**
  * Don't forget the comment
