+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/search/internal/converters/ProductToProductListItemConverter.java
@@ -12,7 +12,7 @@
  *
  ******************************************************************************/
 
-package com.adobe.cq.commerce.core.search.internal;
+package com.adobe.cq.commerce.core.search.internal.converters;
 
 import java.util.Locale;
 import java.util.function.Function;
