+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/search/models/FilterAttributeMetadata.java
@@ -12,7 +12,7 @@
  *
  ******************************************************************************/
 
-package com.adobe.cq.commerce.core.search;
+package com.adobe.cq.commerce.core.search.models;
 
 public interface FilterAttributeMetadata {
