+++ bundles/core/src/main/java/com/adobe/cq/commerce/core/search/models/SearchResultsSet.java
@@ -12,7 +12,7 @@
  *
  ******************************************************************************/
 
-package com.adobe.cq.commerce.core.search;
+package com.adobe.cq.commerce.core.search.models;
 
 import java.util.List;
 import java.util.Map;
