+++ bundles/core/src/test/java/com/adobe/cq/commerce/core/search/internal/AggregationOptionToSearchAggregationOptionConverterTest.java
@@ -23,7 +23,9 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
-import com.adobe.cq.commerce.core.search.SearchAggregationOption;
+import com.adobe.cq.commerce.core.search.internal.converters.AggregationOptionToSearchAggregationOptionConverter;
+import com.adobe.cq.commerce.core.search.internal.models.FilterAttributeMetadataImpl;
+import com.adobe.cq.commerce.core.search.models.SearchAggregationOption;
 import com.adobe.cq.commerce.magento.graphql.AggregationOption;
 
 import static org.assertj.core.api.Assertions.assertThat;
